Creacion del MapComponent y intento de incrustación del mapa en la página de Busqueda.jsx
